+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v4.0.1
 
 - Fix Android Q Easter Egg 'one' text rotation animation [#587](https://github.com/hushenghao/AndroidEasterEggs/issues/587)
+- Fix Android 16 Easter Egg get dessert code crash on Android 8.1 [#597](https://github.com/hushenghao/AndroidEasterEggs/issues/597)
 
 ### v4.0.0 (2025-06-19)